apps/backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
psycopg[binary]==3.3.2
psycopg_pool>=3.2.0
pgvector==0.4.2
pydantic-settings==2.12.0
httpx==0.28.1
google-genai==1.57.0
tenacity>=8.2.0
numpy>=1.26.0

# Redis cache (optional, for production)
redis>=5.0.0

# Database migrations
alembic>=1.13.0
sqlalchemy>=2.0.0

# Observability (metrics)
prometheus_client>=0.19.0

# Auth (JWT + password hashing)
PyJWT>=2.8.0
argon2-cffi>=23.1.0

# Storage (S3/MinIO)
boto3>=1.34.0

# Multipart form handling
python-multipart>=0.0.9

# Background jobs (Redis queue)
rq>=1.16.0

# Document parsing
pypdf>=4.2.0
python-docx>=1.1.0

# Testing dependencies
pytest==9.0.2
pytest-cov==7.0.0
pytest-asyncio==1.3.0

# Tracing (optional, enable with OTEL_ENABLED=1)
# Uncomment to enable OpenTelemetry:
# opentelemetry-api>=1.22.0
# opentelemetry-sdk>=1.22.0
# opentelemetry-instrumentation-fastapi>=0.43b0
